// src/features/infrastructures/constants/enums.ts
export enum InfrastructureStatusEnum {
  New = 'NEW',
  OrderDrawing = 'ORDER_DRAWING',
  ToComplete = 'TO_COMPLETE',
  Completed = 'COMPLETED',
  Returned = 'RETURNED',
  Canceled = 'CANCELED',
  Suspended = 'SUSPENDED',
  ApproveAwaiting = 'APPROVE_AWAITING',
}

export enum BudgetTypeEnum {
  OPEX = 'OPEX',
  CAPEX = 'CAPEX',
}